Avatar saving state and loader + profile refactor
